Extract `Evaluator` from `Watcher`
